requirements/requirements-base.txt
django==3.0.10
djangorestframework-simplejwt==4.4.0
djangorestframework==3.12.1
django-model-utils==4.0.0
djangorestframework-api-key==2.0.0
django-filter==2.4.0
dynaconf==3.1.1
toml==0.10.1
whitenoise==5.2.0
redis==3.5.3
channels==2.4.0
channels_redis==3.1.0
django-silk==4.1.0
dj_database_url==0.5.0
django-redis==4.12.1
drf-rw-serializers==1.0.4
celery==5.0.0
django-cors-headers==3.5.0
